fix(share/booking): open apps not Safari, inline share toast, full month
- app.config: add LSApplicationQueriesSchemes (whatsapp, instagram, tiktok,
  snapchat, fb, twitter). Without these, Linking.canOpenURL() always returns
  false on iOS, so Instagram/TikTok/Snapchat/X share deep-links silently fell
  back to opening the website in Safari instead of the native app. This is the
  root cause of the "share behaves oddly / slow" report
- ShareModal: replace the intrusive OS Alert with a discreet inline
  confirmation banner inside the sheet ("Link kopiert — füg ihn ein…" /
  "{platform} wird geöffnet — Link kopiert"). Sheet stays open so the
  confirmation is visible when returning from the other app
- booking summary: spell the month out in full ("22. August 2026" instead of
  "22. Aug. 2026") to match the events list

+++ game-over-app/app/booking/[eventId]/summary.tsx
@@ -166,7 +166,7 @@ export default function BookingSummaryScreen() {
     const raw = isDraft ? wizardStartDate : ((bookingFlow.event as any)?.start_date || wizardStartDate);
     if (!raw) return null;
     const d = new Date(raw);
-    return isNaN(d.getTime()) ? null : d.toLocaleDateString(language === 'de' ? 'de-DE' : 'en-US', { month: 'short', day: 'numeric', year: 'numeric' });
+    return isNaN(d.getTime()) ? null : d.toLocaleDateString(language === 'de' ? 'de-DE' : 'en-US', { month: 'long', day: 'numeric', year: 'numeric' });
   })();
 
   // Guest count: URL params always take precedence (wizard passes intended group size)
